flesh out tests
